+++ OneBlink.SDK/models/FormStoreSearchMultipleSelectionsArrayFilter.cs
@@ -1,0 +1,13 @@
+using Newtonsoft.Json;
+
+namespace OneBlink.SDK.Model
+{
+    public class FormStoreSearchMultipleSelectionsArrayFilter : IFormStoreInterface
+    {
+        [JsonProperty("$elemMatch")]
+        public FormStoreSearchElementMatchFilter elementMatch
+        {
+            get; set;
+        }
+    }
+}