+++ OneBlink.SDK/models/FormStoreSearchStringArrayComplianceFilter.cs
@@ -1,0 +1,16 @@
+using System;
+
+namespace OneBlink.SDK.Model
+{
+    public class FormStoreSearchStringArrayComplianceFilter : IFormStoreInterface
+    {
+        public FormStoreSearchStringArrayFilter value
+        {
+            get; set;
+        }
+        public FormStoreSearchRegexFilter notes
+        {
+            get; set;
+        }
+    }
+}